Update navigation libraries and simplify navigation set up
