+++ templates/expo-template-tabs/package.json
@@ -16,9 +16,9 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^12.0.0",
-    "@react-navigation/bottom-tabs": "^6.0.4",
+    "@react-navigation/bottom-tabs": "^6.0.5",
     "@react-navigation/native": "^6.0.2",
-    "@react-navigation/native-stack": "^6.0.5",
+    "@react-navigation/native-stack": "^6.1.0",
     "expo": "~42.0.0-alpha.0",
     "expo-asset": "~8.3.2",
     "expo-constants": "~11.0.0",
